# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boiler Service Technician API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: job_type
          in: query
          schema:
            type: string
        - name: technician_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: property_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListItem"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Soft-delete job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
  /jobs/{jobId}/notes:
    get:
      operationId: listJobNotes
      tags: [notes]
      summary: List notes for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobNote"
    post:
      operationId: createJobNote
      tags: [notes]
      summary: Create a job note
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobNote"
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List file attachments
      parameters:
        - name: entity_type
          in: query
          required: true
          schema:
            type: string
        - name: entity_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileAttachment"
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - entity_type
                - entity_id
              properties:
                file:
                  type: string
                  format: binary
                entity_type:
                  type: string
                entity_id:
                  type: string
                  format: uuid
                description:
                  type: string
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileAttachment"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file attachment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /files/{id}/url:
    get:
      operationId: getFileUrl
      tags: [files]
      summary: Get signed download URL for file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUrl"
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, email, full_name, role, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, office_staff, technician]
        phone:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Customer:
      type: object
      required: [id, first_name, last_name, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: ["string", "null"]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address_line1:
          type: ["string", "null"]
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Property:
      type: object
      required: [id, customer_id, address_line1, postcode, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        address_line1:
          type: string
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: string
        property_type:
          type: ["string", "null"]
        occupancy_type:
          type: ["string", "null"]
        access_notes:
          type: ["string", "null"]
        boiler_location:
          type: ["string", "null"]
        flue_location:
          type: ["string", "null"]
        tank_location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        customer_name:
          type: ["string", "null"]
...
    Appliance:
      type: object
      required: [id, property_id, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serial_number:
          type: ["string", "null"]
        boiler_type:
          type: ["string", "null"]
        fuel_type:
          type: ["string", "null"]
        system_type:
          type: ["string", "null"]
        installation_date:
          type: ["string", "null"]
          format: date
        warranty_expiry:
          type: ["string", "null"]
          format: date
        burner_make:
          type: ["string", "null"]
        burner_model:
          type: ["string", "null"]
        nozzle_size:
          type: ["string", "null"]
        pump_pressure:
          type: ["string", "null"]
        controls:
          type: ["string", "null"]
        last_service_date:
          type: ["string", "null"]
          format: date
        next_service_due:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        property_address:
          type: ["string", "null"]
...
    Job:
      type: object
      required: [id, customer_id, property_id, job_type, status, priority, scheduled_date, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        appliance_id:
          type: ["string", "null"]
          format: uuid
        assigned_technician_id:
          type: ["string", "null"]
          format: uuid
        job_type:
          type: string
          enum: [service, breakdown, installation, inspection, follow_up]
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, requires_follow_up]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: ["string", "null"]
        estimated_duration:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobListItem:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            customer_name:
              type: ["string", "null"]
            property_address:
              type: ["string", "null"]
            technician_name:
              type: ["string", "null"]
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            property:
              $ref: "#/components/schemas/Property"
            appliance:
              $ref: "#/components/schemas/Appliance"
            technician:
              $ref: "#/components/schemas/Profile"
            service_record:
              $ref: "#/components/schemas/ServiceRecord"
            breakdown_report:
              $ref: "#/components/schemas/BreakdownReport"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/JobNote"
            files:
              type: array
              items:
                $ref: "#/components/schemas/FileAttachment"
            signatures:
              type: array
              items:
                $ref: "#/components/schemas/Signature"
    CreateJobBody:
      type: object
      required: [customer_id, property_id, job_type, scheduled_date]
      properties:
        customer_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        appliance_id:
          type: string
          format: uuid
        assigned_technician_id:
          type: string
          format: uuid
        job_type:
          type: string
          enum: [service, breakdown, installation, inspection, follow_up]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
        estimated_duration:
          type: integer
        description:
          type: string
        notes:
          type: string
    UpdateJobBody:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        appliance_id:
          type: ["string", "null"]
          format: uuid
        assigned_technician_id:
          type: ["string", "null"]
          format: uuid
        job_type:
          type: string
          enum: [service, breakdown, installation, inspection, follow_up]
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, requires_follow_up]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: ["string", "null"]
        estimated_duration:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ServiceRecord:
      type: object
      required: [id, job_id, technician_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        arrival_time:
          type: ["string", "null"]
          format: date-time
        departure_time:
          type: ["string", "null"]
          format: date-time
        visual_inspection:
          type: ["string", "null"]
        appliance_condition:
          type: ["string", "null"]
        flue_inspection:
          type: ["string", "null"]
        combustion_co2:
          type: ["string", "null"]
        combustion_co:
          type: ["string", "null"]
        combustion_o2:
          type: ["string", "null"]
        combustion_temp:
          type: ["string", "null"]
        combustion_efficiency:
          type: ["string", "null"]
        smoke_test:
          type: ["string", "null"]
        smoke_number:
          type: ["string", "null"]
        burner_cleaned:
          type: boolean
        heat_exchanger_cleaned:
          type: boolean
        nozzle_checked:
          type: boolean
        nozzle_replaced:
          type: boolean
        nozzle_size_fitted:
          type: ["string", "null"]
        electrodes_checked:
          type: boolean
        electrodes_replaced:
          type: boolean
        filter_checked:
          type: boolean
        filter_cleaned:
          type: boolean
        filter_replaced:
          type: boolean
        oil_line_checked:
          type: boolean
        fire_valve_checked:
          type: boolean
        seals_gaskets_checked:
          type: boolean
        seals_gaskets_replaced:
          type: boolean
        controls_checked:
          type: boolean
        thermostat_checked:
          type: boolean
        safety_devices_checked:
          type: boolean
        safety_devices_notes:
          type: ["string", "null"]
        leaks_found:
          type: boolean
        leaks_details:
          type: ["string", "null"]
        defects_found:
          type: boolean
        defects_details:
          type: ["string", "null"]
        advisories:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        work_completed:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        follow_up_required:
          type: boolean
        follow_up_notes:
          type: ["string", "null"]
        next_service_due:
          type: ["string", "null"]
          format: date
        additional_notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BreakdownReport:
      type: object
      required: [id, job_id, technician_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        reported_fault:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        diagnostics_performed:
          type: ["string", "null"]
        findings:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        temporary_fix:
          type: ["string", "null"]
        permanent_fix:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        return_visit_required:
          type: boolean
        return_visit_notes:
          type: ["string", "null"]
        additional_notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    JobNote:
      type: object
      required: [id, job_id, author_id, content, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        content:
          type: string
        author_name:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateJobNoteBody:
      type: object
      required: [content]
      properties:
        content:
          type: string
    FileAttachment:
      type: object
      required: [id, file_name, file_type, storage_path, entity_type, entity_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
        file_type:
          type: string
        file_size:
          type: ["integer", "null"]
        storage_path:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        uploaded_by:
          type: ["string", "null"]
          format: uuid
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        signed_url:
          type: ["string", "null"]
    FileUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string
    Signature:
      type: object
      required: [id, job_id, signer_type, signer_name, storage_path, signed_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        signer_type:
          type: string
          enum: [technician, customer]
        signer_name:
          type: string
        storage_path:
          type: string
        signed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        signed_url:
          type: ["string", "null"]
...