# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Boiler Service Technician API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateJobBodyJobType } from "./updateJobBodyJobType";
import type { UpdateJobBodyPriority } from "./updateJobBodyPriority";
import type { UpdateJobBodyStatus } from "./updateJobBodyStatus";

export interface UpdateJobBody {
  customer_id?: string;
  property_id?: string;
  /** @nullable */
  appliance_id?: string | null;
  /** @nullable */
  assigned_technician_id?: string | null;
  job_type?: UpdateJobBodyJobType;
  status?: UpdateJobBodyStatus;
  priority?: UpdateJobBodyPriority;
  scheduled_date?: Date;
  /** @nullable */
  scheduled_time?: string | null;
  /** @nullable */
  estimated_duration?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
}